Move additional channel opening tests to `channel_open_tests.rs`
`functional_tests.rs` is a rather large beast, but with this move
we take it down to under 10k LoC.
